refac(args): make args local and activate tests
